dashboard/requirements.txt
streamlit
psycopg2-binary
pandas
matplotlib
wordcloud
python-dotenv